backend_python/requirements.txt

fastapi
uvicorn
python-dotenv
httpx


ollama
mem0ai

qdrant-client 
aiosqlite
mcp
PyYAML

ruff
pytest
pytest-asyncio

numpy
tiktoken

watchfiles

Pillow

python-telegram-bot
discord.py

neo4j
apscheduler

croniter
PyJWT
websockets
